blockchain: move constructor functions to 'constructors.ts'
