+++ packages/blockchain/src/index.ts
@@ -1,5 +1,6 @@
 export { Blockchain } from './blockchain.js'
 export { CasperConsensus, CliqueConsensus, EthashConsensus } from './consensus/index.js'
+export * from './constructors.js'
 export {
   DBOp,
   DBSaveLookups,
